Moved TokenRepositories to their own package
