+++ src/main/java/ch/wisv/areafiftylan/controller/AuthenticationController.java
@@ -7,8 +7,8 @@ import ch.wisv.areafiftylan.security.PasswordResetToken;
 import ch.wisv.areafiftylan.security.VerificationToken;
 import ch.wisv.areafiftylan.service.MailService;
 import ch.wisv.areafiftylan.service.UserService;
-import ch.wisv.areafiftylan.service.repository.PasswordResetTokenRepository;
-import ch.wisv.areafiftylan.service.repository.VerificationTokenRepository;
+import ch.wisv.areafiftylan.service.repository.token.PasswordResetTokenRepository;
+import ch.wisv.areafiftylan.service.repository.token.VerificationTokenRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
