+++ src/main/java/ch/wisv/areafiftylan/service/UserServiceImpl.java
@@ -6,9 +6,9 @@ import ch.wisv.areafiftylan.model.Profile;
 import ch.wisv.areafiftylan.model.User;
 import ch.wisv.areafiftylan.security.PasswordResetToken;
 import ch.wisv.areafiftylan.security.VerificationToken;
-import ch.wisv.areafiftylan.service.repository.PasswordResetTokenRepository;
+import ch.wisv.areafiftylan.service.repository.token.PasswordResetTokenRepository;
 import ch.wisv.areafiftylan.service.repository.UserRepository;
-import ch.wisv.areafiftylan.service.repository.VerificationTokenRepository;
+import ch.wisv.areafiftylan.service.repository.token.VerificationTokenRepository;
 import com.google.common.base.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
