+++ src/main/java/ch/wisv/areafiftylan/service/repository/token/PasswordResetTokenRepository.java
@@ -1,4 +1,4 @@
-package ch.wisv.areafiftylan.service.repository;
+package ch.wisv.areafiftylan.service.repository.token;
 
 import ch.wisv.areafiftylan.security.PasswordResetToken;
 import org.springframework.data.jpa.repository.JpaRepository;
