+++ src/main/java/ch/wisv/areafiftylan/service/repository/token/VerificationTokenRepository.java
@@ -1,4 +1,4 @@
-package ch.wisv.areafiftylan.service.repository;
+package ch.wisv.areafiftylan.service.repository.token;
 
 import ch.wisv.areafiftylan.security.VerificationToken;
 import org.springframework.data.jpa.repository.JpaRepository;
